+++ CHANGELOG.md
@@ -17,6 +17,14 @@
     Click to see more.
   </summary>
 
+### :rocket: New Feature
+
+- `ricos-editor`
+  - [#3068](https://github.com/wix/ricos/pull/3068) **beta** Editor Commands API support for Notes
+
+</details>
+<hr/>
+
 </details>
 <hr/>